Split powerline/regular otf, add prepare script
The powerline mod script from nerd-fonts seems to ruin a lot of
the preparation of the regular font, so its been split out from
the regular version of the font and will not be tested against the
fontbakery guidelines.

A script was also added to help with preparing the font for upload.