Add 56 glyphs, fix otf name, add googlefonts ci
This adds the remaining 56 glyphs necessary for passing the googlefonts
validator in fontbakery.

The OTF names were updated to match their expected format of
FontFamily-Type.otf.

The googlefonts fontbakery check was added as a separate ci step. It was
included separately from the universal check to allow for easier viewing
of what is right/wrong in the build.